Refactor NmeaParser
