+++ app/CMakeLists.txt
@@ -35,6 +35,7 @@ set(MM_SRCS
     position/providers/networkpositionprovider.cpp
     position/providers/positionprovidersmodel.cpp
     position/providers/simulatedpositionprovider.cpp
+    position/providers/nmeaparser.cpp
     position/tracking/abstracttrackingbackend.cpp
     position/tracking/internaltrackingbackend.cpp
     position/tracking/positiontrackinghighlight.cpp
@@ -130,6 +131,7 @@ set(MM_HDRS
     position/providers/networkpositionprovider.h
     position/providers/positionprovidersmodel.h
     position/providers/simulatedpositionprovider.h
+    position/providers/nmeaparser.h
     position/tracking/abstracttrackingbackend.h
     position/tracking/internaltrackingbackend.h
     position/tracking/positiontrackinghighlight.h
