+++ app/position/providers/bluetoothpositionprovider.h
@@ -10,29 +10,12 @@
 #ifndef BLUETOOTHPOSITIONPROVIDER_H
 #define BLUETOOTHPOSITIONPROVIDER_H
 
-#include "abstractpositionprovider.h"
-
-#include "qgsnmeaconnection.h"
-
 #include <QBluetoothSocket>
 #include <QBluetoothLocalDevice>
 #include <QTimer>
 
-/**
- * NmeaParser is a big hack how to reuse QGIS NmeaConnection function in order to (a) keep ownership of bluetooth
- * socket, (b) do not have multiple unique_ptrs holding the same pointer and to avoid some possible crashes.
- *
- * Note: This way of reusing makes the parser highly dependent on QgsNmeaConnection class and any change inside the class
- * can lead to misbehavior's. See implementation of QgsNmeaConnection and QgsGpsConnection for more details.
- */
-class NmeaParser : public QgsNmeaConnection
-{
-  public:
-    NmeaParser();
-
-    // Takes nmea string and returns gps position
-    QgsGpsInformation parseNmeaString( const QString &nmeaString );
-};
+#include "abstractpositionprovider.h"
+#include "nmeaparser.h"
 
 /**
  * BluetoothPositionProvider initiates connection to bluetooth device
@@ -51,7 +34,7 @@ class BluetoothPositionProvider : public AbstractPositionProvider
     };
 
   public:
-    BluetoothPositionProvider( const QString &addr, const QString &name, PositionTransformer &positionTransformer, QObject *parent = nullptr );
+    BluetoothPositionProvider( const QString &addr, const QString &name, QObject *parent = nullptr );
     ~BluetoothPositionProvider() override;
 
     void startUpdates() override;
